Use floating point for units

+++ Source/GraphicsModule.cpp
@@ -29,8 +29,8 @@ struct GT_Texture
 
 /* ====== GLOBALS ====== */
 GraphicsModule g_graphicsModule;
-s32 g_unitX = 0;
-s32 g_unitY = 0;
+f32 g_unitX = 0;
+f32 g_unitY = 0;
 
 /* ====== METHODS ====== */
 b32 GraphicsModule::StartUp(SDL_Renderer* pRenderer, s32 width, s32 height)
@@ -38,8 +38,8 @@ b32 GraphicsModule::StartUp(SDL_Renderer* pRenderer, s32 width, s32 height)
     // Defaults
     m_screenWidth = width;
     m_screenHeight = height;
-    g_unitX = m_screenWidth / UNIT_SCREEN_WIDTH;
-    g_unitY = m_screenHeight / UNIT_SCREEN_HEIGHT;
+    g_unitX = m_screenWidth / (f32)UNIT_SCREEN_WIDTH;
+    g_unitY = m_screenHeight / (f32)UNIT_SCREEN_HEIGHT;
     m_pRenderer = pRenderer;
 
     // Allocate textures
